app/src/main/java/hristostefanov/creditscoredemo/presentation/MainViewModel.kt
package hristostefanov.creditscoredemo.presentation

import androidx.lifecycle.ViewModel
import dagger.hilt.android.lifecycle.HiltViewModel
import hristostefanov.creditscoredemo.business.dependencies.CreditScoreRepository
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.StateFlow
import javax.inject.Inject

@HiltViewModel
class MainViewModel @Inject constructor(creditScoreRepository: CreditScoreRepository) :
    ViewModel() {
    private val _viewState = MutableStateFlow(MainViewState())
    val viewState: StateFlow<MainViewState> = _viewState
}